+++ CHANGELOG.md
@@ -34,6 +34,9 @@ Please document your changes in this format:
 ### Removed
 - Remove calendar subscribe button from group wall @amengsk [#2451]
 
+### Added
+- Choice of default view (wall or activities) on place page @brnsolikyl @nicksellen @larzon83 [#2444]
+
 ## [9.5.1] - 2021-11-08
 ### Changed
 - increase android API level target from 29 to 30 @nicksellen